refactor(hickey,lowy): close canvas/terminal boundary and trust TerminalKey
- Placement consumes the canonical `TerminalKey` projection (no string-only
  group field on the snapshot). When the new cwd has no resolved git, it
  inherits a sibling's git and projects through `terminalKey()` directly,
  so `terminalKey` is the single source of group derivation.
- Canvas no longer knows about repo grouping: `autoArrange` takes a
  `arrange(current)` callback. App.tsx owns the call to `arrangeRepoIslands`
  and feeds it canonical `key.group` values.
- Drop the `autoArrange.ts` re-export wrapper; tests fold into the
  placement test file.
- `useTerminalCrud` no longer imports from `canvas/`. The composition root
  (`useTerminals`) injects a `computeInitialLayout` callback so CRUD stays
  agnostic to the placement policy in force.
